move tests to subsets controller test class
